+++ CHANGES.md
@@ -26,4 +26,5 @@
 ### Released 9/10/19
 -   Add Sort Button #30
 -   Add props to displayName override function #37
+-   Standardize isDeletable, isCreatable, isEditable & others #35
 -   Add sortable prop to schema #38